requirements.txt
selenium==4.25.0
python-dotenv==1.0.1
fastapi
uvicorn[standard]
webdriver-manager